Move stats code
