+++ src/backend/compare/charts.ts
@@ -12,7 +12,7 @@ import {
   BoxPlotController,
   BoxAndWiskers
 } from '@sgratzl/chartjs-chart-boxplot';
-import { medianUnsorted } from '../../stats.js';
+import { medianUnsorted } from '../../shared/stats.js';
 import { robustPath, siteAesthetics } from '../util.js';
 
 const fullyTransparent = 'rgba(0, 0, 0, 0)';
