+++ src/backend/db/db.ts
@@ -22,7 +22,7 @@ import {
 import { robustPath, TotalCriterion } from '../util.js';
 import { assert } from '../logging.js';
 import { simplifyCmdline } from '../../views/util.js';
-import type { SummaryStatistics } from '../../stats.js';
+import type { SummaryStatistics } from '../../shared/stats.js';
 import { BatchingTimelineUpdater } from '../timeline/timeline-calc.js';
 import type {
   Baseline,
