+++ src/shared/stats.ts
